+++ src/cl_parser/cl_parser.h
@@ -1,8 +1,8 @@
 #pragma once
 
 struct ProgramConfig {
-    long thread_count{};
-    long task_count{};
+    int thread_count{};
+    int task_count{};
 };
 
 ProgramConfig ParseCl(int argc, const char* argv[]);
